src/main/java/csci310/servlets/LogOutServlet.java
package csci310.servlets;

import csci310.dao.UserDAO;
import csci310.model.User;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;
import java.sql.SQLException;
import java.util.ArrayList;

public class LogOutServlet extends HttpServlet {

    // log the user out and redirect to /login
    @Override
    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
        request.getSession().invalidate();
        response.sendRedirect("/login");
    }

}
